feat(wasm): execute memory, calls, globals, instantiation
Complete Cut 4: function calls and tail calls, global get/set,
linear-memory load/store with OOB traps, data segments, and
instantiation with start functions.

- exec.rs: explicit-frame machine gains Memory (64 KiB pages,
  grow capped at 2^16 pages) and LE loads/stores of every width
  with sign/zero extension; return_call replaces the top frame in
  place at the same operand-stack base so tail recursion does not
  consume the depth budget.
- binary.rs: memory limits decode as u64 LEB; memargs are the
  post-memory64 form (single flags byte with alignment exponent
  and memory-index bit, u64 offset), matching wabt's encoding of
  the assert_invalid fixtures.
- valid.rs: tail calls type with subsumption (a callee may yield
  (ref null $t) where funcref was declared); memory32 limits are
  bounded at 2^16 pages for defined and imported memories; memarg
  offsets above 2^32-1 are "offset out of range".
- wasmtest: quote-based (.wat) modules report pending instead of
  being misjudged through the binary decoder.
- .notes/wasm-plan.md: Cut 4 gate results and deferral taxonomy.

Gate: 0 failures across the function/call/memory/load/store/data/
tail-call files, with two written Cut-5 spillovers (uninitialized
typed local; post-memory.init load) and imports/register pending a
follow-up host-import resolver.

+++ crates/wasm/src/instr.rs
@@ -235,12 +235,12 @@ pub enum Instr {
     Load {
         op: LoadOp,
         align: u32,
-        offset: u32,
+        offset: u64,
     },
     Store {
         op: StoreOp,
         align: u32,
-        offset: u32,
+        offset: u64,
     },
     MemorySize,
     MemoryGrow,
